Add files via upload
I've uploaded a new PNG file for the small logo icon, the one that you see in the tab. The one that we use now seems to be a little distorted, so I made a new one, more sqer-like shape. I did't know where to put it, so I've put it inside the "image" folder. also I've put there before, Ravit's site image. All this I wrote in En! took me only 5-6 min